# openapi.yaml
openapi: 3.0.3
info:
    title: MechHub Miniback API
    version: 1.0.0
    description: Standalone backend for MechHub. Authentication is session-based and uses the `mechhub.sid` cookie.
servers:
    - url: http://localhost:3001
      description: Local development server
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: mechhub.sid
    schemas:
...
        ErrorResponse:
            type: object
            required:
                - message
            properties:
                message:
                    type: string
                    example: Not authenticated
...
        FileAsset:
            type: object
            required:
                - id
                - fileName
                - url
                - mimeType
                - sizeBytes
                - purpose
                - createdAt
            properties:
                id:
                    type: integer
                    example: 21
                fileName:
                    type: string
                    example: assignment-week-3.pdf
                url:
                    type: string
                    example: /uploads/files/21
                mimeType:
                    type: string
                    example: application/pdf
                sizeBytes:
                    type: integer
                    example: 184291
                width:
                    type: integer
                    nullable: true
                    example: 1536
                height:
                    type: integer
                    nullable: true
                    example: 1024
                kind:
                    type: string
                    enum: [image, document, file]
                    example: image
                purpose:
                    type: string
                    enum:
                        [assignment_attachment, submission_attachment, solochat]
                    example: assignment_attachment
                createdAt:
                    type: string
                    format: date-time
                    example: 2026-03-25T09:31:00.456Z
...
paths:
...
    /uploads/files:
        post:
            tags:
                - Uploads
            summary: Upload a reusable file attachment
            operationId: uploadFile
            security:
                - cookieAuth: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required:
                                - file
                            properties:
                                file:
                                    type: string
                                    format: binary
                purpose:
                    type: string
                    enum:
                        [assignment_attachment, submission_attachment, solochat]
                    example: solochat
            responses:
                "201":
                    description: File uploaded
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/FileAsset"
                "400":
                    description: Unsupported or oversized file
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /uploads/files/{fileId}:
        get:
            tags:
                - Uploads
            summary: Download a previously uploaded file attachment
            operationId: getUploadedFile
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: fileId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Uploaded file
                    content:
                        application/octet-stream:
                            schema:
                                type: string
                                format: binary
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: File not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        delete:
            tags:
                - Uploads
            summary: Delete an unattached uploaded file
            operationId: deleteUploadedFile
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: fileId
                  required: true
                  schema:
                      type: integer
            responses:
                "204":
                    description: Uploaded file deleted
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: File not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: File is already attached to an assignment or submission
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...